refactor: move FunctionProvider to feel module
