+++ expression-language/src/main/java/io/camunda/zeebe/el/impl/FeelExpressionLanguage.java
@@ -14,9 +14,9 @@ import io.camunda.zeebe.el.EvaluationResult;
 import io.camunda.zeebe.el.Expression;
 import io.camunda.zeebe.el.ExpressionLanguage;
 import io.camunda.zeebe.el.impl.feel.FeelEvaluationResult;
-import io.camunda.zeebe.el.impl.feel.FeelFunctionProvider;
 import io.camunda.zeebe.el.impl.feel.FeelToMessagePackTransformer;
 import io.camunda.zeebe.el.impl.feel.FeelVariableContext;
+import io.camunda.zeebe.feel.impl.FeelFunctionProvider;
 import io.camunda.zeebe.feel.impl.MessagePackValueMapper;
 import io.camunda.zeebe.util.sched.clock.ActorClock;
 import java.util.regex.Pattern;
