+++ feel/src/main/scala/io/camunda/zeebe/feel/impl/FeelFunctionProvider.scala
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.el.impl.feel
+package io.camunda.zeebe.feel.impl
 
 import org.camunda.feel.context.FunctionProvider
 import org.camunda.feel.syntaxtree._
